Removed most of the warnings. Had mostly data type conversion leading to loss of data warnings.

+++ AMMPED-UP/Header Files/Button.h
@@ -57,8 +57,8 @@ public:
 
 	bool isMouseOver(sf::RenderWindow& window)
 	{
-		float mouseX = sf::Mouse::getPosition(window).x;
-		float mouseY = sf::Mouse::getPosition(window).y;
+		float mouseX = static_cast<float>(sf::Mouse::getPosition(window).x);
+		float mouseY = static_cast<float>(sf::Mouse::getPosition(window).y);
 
 		float btnPosX = button.getPosition().x;
 		float btnPosY = button.getPosition().y;
